requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
blinker==1.4
cherrypy==3.8.0
discord.py==0.8.0
itsdangerous==0.24
oauthlib==1.0.3           # via requests-oauthlib
requests-oauthlib==0.5.0
requests==2.8.1           # via discord.py, requests-oauthlib
ws4py==0.3.4              # via discord.py
